start moving the search box into the omni-prompt and place it under a map icon between back/forward
