+++ src/renderer/app.css
@@ -312,8 +312,7 @@ input {
   z-index: 10;
   position: fixed;
   top: 52px;
-  right: 5%;
-  margin-left: -239px;
+  left: 12px;
 }
 
 .Omnibox .ListMenu__item .Actions {
